Build container and push to github packages
Adds a docker build and push to get a version of the package up in github packages. For now it'll build and push a docker image once we merge to main. In reality we'd want image for every push on a PR to also build an image so that we can deploy to a staging env to test branches, but keeping things simple for now.